renamed modules name (fix issue related to refactor modules name to lowercase characters)
