+++ applications/smart-distancing/libs/core.py
@@ -16,10 +16,10 @@ class Distancing:
         self.tracker = CentroidTracker(
             max_disappeared=int(self.config.get_section_dict("PostProcessor")["MaxTrackFrame"]))
         if self.device == 'Jetson':
-            from libs.detectors.jetson.Detector import Detector
+            from libs.detectors.jetson.detector import Detector
             self.detector = Detector(self.config)
         elif self.device == 'EdgeTPU':
-            from libs.detectors.edgetpu.Detector import Detector
+            from libs.detectors.edgetpu.detector import Detector
             self.detector = Detector(self.config)
         elif self.device == 'Dummy':
             self.detector = None
